Various fixes after changes to instanced mesh management
